Add container notes for Gnome 40
Ubuntu 21.10 already works by default inside this container implementation.
However some additional steps might be necessary to run this succesfully on
older versions of Docker. Add some notes to clarify this.

Also fix the mentioning of debug messages. Not all messages are relayed to
the host console for some reason.