Update workarounds for outdated dependencies in Ubuntu runner
With this commit an unsupported package version of libadwaita is
installed in the runner. As a result, builds from this runner are not
really guaranteed to be free from error since the build system is
untested/unsupported.

The entire workaround for libadwaita should be removed eventually when a
newer Ubuntu version is available.