Image preview widget in popup menu
This commit replaces the image preview in the sub-menus with a single
preview widget directly in the popup menu of the extension. When a
history element is hovered or focused with the keyboard, the corresponding image
will be displayed in the preview widget.

Preview events are debounced by a few milliseconds to prevent
unnecessary image loads.