Remove CurrentImageElement and cached history elements
This change is done to simplify the UI code and remove inconsistencies
when a wallpaper is changed without using this extension. I.e., the
current wallpaper element was not aware of external wallpaper changes.

Additionally, this commit removes the element cache for HistorySection
which likely isn't necessary now since the preview images are not part
of the history elements anymore.