Clear the entire history
The first history element was kept around when the history was cleared
so that the current wallpaper element could be displayed. With the
removal of the this element, there is no need to keep the first entry in
the history.

Note that we do not delete the file of the first entry so the wallpaper
remains on the disk and can be used until the next wallpaper is
requested.